One More Step Towards ORM

+++ project_spring_boot_backend/src/main/java/com/app/entities/OrderTableId.java
@@ -24,10 +24,10 @@ public class OrderTableId implements Serializable{
 	private Long orderId; 
 	
 	 @Column(name = "rest_id")
-	 private Integer restId;
+	 private Long restId;
 
 	 @Column(name = "item_id")
-	 private Integer itemId;
+	 private Long itemId;
 	 
 	 
 	 @Override
